Writing direcly to JSON files instead of gviz_api
The gviz_api import writes the entire datatable to the JSON file each
time.
I removed it and write to JSON files direcly. Adding a datapoint is now
done by opening the JSON file, seeking 3 bytes before the end of the
file and appending the new data. Should run a lot lighter.